refactor: 에러 레이아웃 메시지, 푸터 에러 메시지 상수화
